Rename Automation library to GitAutomation
