+++ Microsoft.DotNet.DockerTools.slnx
@@ -7,8 +7,8 @@
   <Project Path="eng/src/file-pusher/file-pusher.csproj" />
   <Project Path="eng/src/yaml-updater/yaml-updater.csproj" />
   <Project Path="src/ImageBuilder/Microsoft.DotNet.ImageBuilder.csproj" />
-  <Project Path="src/Automation/Microsoft.DotNet.Automation.csproj" />
-  <Project Path="src/Automation.Tests/Microsoft.DotNet.Automation.Tests.csproj" />
+  <Project Path="src/GitAutomation/Microsoft.DotNet.GitAutomation.csproj" />
+  <Project Path="src/GitAutomation.Tests/Microsoft.DotNet.GitAutomation.Tests.csproj" />
   <Project Path="src/ImageBuilder.Models/Microsoft.DotNet.ImageBuilder.Models.csproj" />
   <Project Path="src/ImageBuilder.Tests/Microsoft.DotNet.ImageBuilder.Tests.csproj" />
 </Solution>
