+++ src/GitAutomation.Tests/ProcessRunnerExtensionsTests.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.DotNet.Automation.Tests;
+namespace Microsoft.DotNet.GitAutomation.Tests;
 
 [TestClass]
 public sealed class ProcessRunnerExtensionsTests
