+++ src/GitAutomation/GitWorkspace.cs
@@ -4,7 +4,7 @@
 
 using Microsoft.Extensions.Logging;
 
-namespace Microsoft.DotNet.Automation;
+namespace Microsoft.DotNet.GitAutomation;
 
 internal sealed class GitWorkspace(string directory, ILogger logger) : IDisposable
 {
