+++ src/GitAutomation/Microsoft.DotNet.GitAutomation.csproj
@@ -28,7 +28,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <InternalsVisibleTo Include="Microsoft.DotNet.Automation.Tests" />
+    <InternalsVisibleTo Include="Microsoft.DotNet.GitAutomation.Tests" />
   </ItemGroup>
 
 </Project>
